refactor: move config loading + validation into the mediator-config crate (simplification T18, part b)
Completes T18. Following part a (raw schema), this moves the two remaining
config concerns out of the mediator binary into the shared, dependency-light
`affinidi-messaging-mediator-config` crate so the mediator and (next, T19) the
mediator-setup wizard share one implementation:

- Config loading: `read_config_file` + `apply_env_overrides` (+ the env_override
  macros) → crate `env` module. Because the crate can't use the server-tier
  `MediatorError`, a lean `ConfigError` (thiserror) is introduced; the mediator
  maps it back to `MediatorError::ConfigError` at the single call site in
  `config::init`.
- Boot validation: the pure check helpers (DID syntax, JWT-expiry ordering, TLS
  file presence, and the four legal-but-suspicious-combo warnings) → crate
  `validate` module, with their unit tests. The relay warning is decoupled from
  the binary's `authz` module — `grants(.., Capability::SendForwarded)` becomes
  the direct `MediatorACLSet::get_send_forwarded().0` bit accessor — so the
  schema crate needs only the lean ACL types it already had. The mediator's
  `validate_config(&Config)` stays as a thin orchestrator over the crate checks
  (maps errors to MediatorError, logs warnings).

The crate stays off the server stack (new deps toml/thiserror/tracing are all
lean; verified via cargo tree). `read_file_lines` exists in both crates (the
mediator keeps a private copy for its runtime document/DID helpers).

Behaviour-identical. Verification: config crate 11 tests (golden + 10 moved
validation tests); mediator lib 134; test-mediator e2e all green; clippy clean
on both feature sets; -D warnings build clean. No mediator-common change.
mediator 0.15.41, mediator-config 0.1.1.
