+++ crates/messaging/affinidi-messaging-mediator/affinidi-messaging-mediator-config/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ## 12th June 2026
 
+### 0.1.1 — Config loading + validation (simplification T18, part b)
+
+- Adds the `env` module (config-file reading + env-var overrides:
+  `read_config_file`, `apply_env_overrides`) and the `validate` module (the pure
+  boot-time invariant checks: DID syntax, JWT-expiry ordering, TLS file presence,
+  and the legal-but-suspicious-combo warnings), both moved out of the mediator
+  binary. The mediator now re-uses them; its `validate_config(&Config)` is a thin
+  orchestrator that maps results to `MediatorError` and logs warnings.
+- New lean `ConfigError` (thiserror) for `read_config_file` instead of the
+  mediator's server-tier `MediatorError`. The relay warning is decoupled from the
+  mediator's `authz` module (uses the `MediatorACLSet` SEND_FORWARDED bit
+  accessor directly), so validation needs only the lean ACL types. The crate
+  stays off the server stack (added deps: toml, thiserror, tracing — all lean).
+
 ### 0.1.0 — Initial release (simplification T18, part a)
 
 - New crate holding the mediator's raw TOML configuration **schema** — the
